VoiceMind-Meeting-System/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
mysql-connector-python==8.2.0
openai==1.3.5
pydantic==2.5.0
numpy==1.24.3
scipy==1.11.4
